# deploy/api-tts-v1/openapi.yaml
openapi: 3.0.1
info:
  title: CapCut TTS Wrapper API
  description: Backend-only CapCut wrapper for text and DOCX to MP3 workflows.
  version: 2.2.0
servers:
  - url: http://localhost:8080
    description: Local development server
paths:
...
  /v1/synthesize/docx/jobs/{jobId}:
    get:
      summary: Get async DOCX synthesis job status
      tags:
        - TTS
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Job status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocxJob'
        '404':
          description: Job was not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /v1/synthesize/docx/jobs/{jobId}/download:
    get:
      summary: Download a completed async DOCX synthesis MP3
      tags:
        - TTS
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: MP3 audio
          content:
            audio/mpeg:
              schema:
                type: string
                format: binary
        '404':
          description: Job was not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Job is not ready for download
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          example: Failed to synthesize audio
    DocxJob:
      type: object
      properties:
        jobId:
          type: string
          example: aba39c2cf931bdd77d9df37e4d8a975b
        status:
          type: string
          enum:
            - pending
            - running
            - completed
            - failed
        fileName:
          type: string
          example: output.mp3
        totalChunks:
          type: integer
          example: 1492
        completedChunks:
          type: integer
          example: 42
        progress:
          type: number
          example: 0.028
        progressPercent:
          type: number
          example: 2.8
        active:
          type: boolean
        downloadReady:
          type: boolean
        contentType:
          type: string
          example: audio/mpeg
        contentLength:
          type: string
          example: '1736685'
        errorMessage:
          type: string
        statusUrl:
          type: string
          example: /v1/synthesize/docx/jobs/aba39c2cf931bdd77d9df37e4d8a975b
        downloadUrl:
          type: string
          example: /v1/synthesize/docx/jobs/aba39c2cf931bdd77d9df37e4d8a975b/download
        resumeUrl:
          type: string
          example: /v1/synthesize/docx/jobs/aba39c2cf931bdd77d9df37e4d8a975b/resume
...